FIXED THE BUG
Turns out comparing Strings like (String == String) is a bad idea...
Strings are compared by reference, so if the objects aren't referring to
the same place in memory, it returns false.

The Strings are now being compared with the following String method:
string1.equals(string2)

This accomplishes the desired result.

I have tested decrease item through multiple print out and it IS indeed
decrease the stock field in the Tool object passed.