Added a Randomized Sleep Length in SimpleThread
Answer for task 1:

On my hardware, the output would not get stuck between threads. To cause
it, I had to lengthen the argument in the sleep() call by a random
amount between 0 and 50. This created enough variation for the scheduler
to preempt a thread in the middle of a key computation, in this case the
increment of +1 to the counter in Resource. For instance, if one thread
loaded counter into a register, incremented it, and is then
preempted before it can store the new value in counter, another thread
could begin with a counter that has not been updated. If this other
thread continues past the store point, counter will have the same value
as the one to be stored once the original thread is given a chance to
run again.